requirements.txt
gradio>=4.0.0
torch>=2.1
timm>=0.9.0
albumentations>=1.3.1
opencv-python-headless>=4.8.0
pytorch-grad-cam>=1.4.8
numpy
pandas
Pillow
